all: Fix implicit floating point to integer conversions.
These are found when using building with -Wfloat-conversion.

+++ py/binary.c
@@ -419,10 +419,10 @@ void mp_binary_set_val_array_from_int(char typecode, void *p, size_t index, mp_i
         #endif
         #if MICROPY_PY_BUILTINS_FLOAT
         case 'f':
-            ((float *)p)[index] = val;
+            ((float *)p)[index] = (float)val;
             break;
         case 'd':
-            ((double *)p)[index] = val;
+            ((double *)p)[index] = (double)val;
             break;
         #endif
         // Extension to CPython: array of pointers
